loot jump, enemy spawn, damage,defence,shot

+++ Assets/Scripts/Game/Component/Stat/FarmStats.cs
@@ -5,7 +5,7 @@ namespace Game.Component
     public struct FarmStats
     {
         public CultureType CurrentCulture;
-        public float CultureCoins;
+        public int CultureCoins;
         
         public float GrowthSpeedK;
         public float MultChance;
